Fixing the index for the AppShell
I still need to fix the indexFiltering, like I did it with maps
