+++ assets/css/index.css
@@ -1,42 +1,8 @@
-body {
-	background-color: #000;
-	font-family: Verdana, Tahoma, sans-serif;
-	font-size: 100%;
-	margin: 0;
-}
-
 #shell{
 	display: none;
 }
 
-img#tx {
-	display: inline-block;
-	padding-top: 4px;
-	padding-left: 4px;
-}
-
-img#tx:hover {
-	opacity: 1;
-	transform: scale(1.05);
-}
-
 /* -header- */
-header {
-	width: 100%;
-	height: 64px;
-	position: fixed;
-	background-color: #000;
-	z-index: 1;
-}
-
-h1.logo {
-	font: normal normal bold 1em sans-serif;
-	color: #fff;
-	text-align: center;
-	display: inline-block;
-	margin: 0;
-	width: 0;
-}
 
 #contact {
 	cursor: pointer;
@@ -106,13 +72,6 @@ h1.logo {
 
 /* Mobile */
 @media only screen and (max-width: 756px) {
-	body {
-		background-color: #404040;
-		font-family: Verdana, Tahoma, sans-serif;
-		font-size: 100%;
-		margin: 0;
-	}
-
 	header .menu a {
 		display: none;
 	}
@@ -135,15 +94,6 @@ h1.logo {
 		margin-top: 3px;
 		margin-right: 8px;
 		overflow: visible;
-	}
-
-	h1.logo {
-		font: normal normal bold 1em sans-serif;
-		color: #fff;
-		text-align: center;
-		display: inline-block;
-		margin: 0;
-		width: 60px;
 	}
 	
 	#oi {
@@ -173,13 +123,6 @@ h1.logo {
 
 /*CSS file for the Homepage (Resolution 1024x768 (Viewsize 1010x660) and up)*/
 @media only screen and (min-width: 756px) {
-	body {
-		background-color: #000;
-		font-family: Verdana, Tahoma, sans-serif;
-		font-size: 100%;
-		margin: 0;
-	}
-
 	header .menu a {
 		display: inline-block;
 		color: #fff;
@@ -229,15 +172,6 @@ h1.logo {
 @media only screen and (min-width: 1010px) {
 	#oc {
 		width: 1010px;
-	}
-
-	h1.logo {
-		font: normal normal bold 2.7em sans-serif; /*Verdana, Tahoma,*/
-		color: #fff;
-		text-align: center;
-		display: inline-block;
-		margin: 0;
-		width: auto;
 	}
 }
