+++ assets/css/map.css
@@ -3,15 +3,6 @@
 @import "mapmenucontrol.css";
 @import "mapfeaturetooltip.css";
 
-html,
-body {
-	margin: 0;
-	height: 100%;
-	background-color: #000;
-	font-family: Verdana, Tahoma, sans-serif;
-	font-size: 100%;
-}
-
 .map-html {
 	overflow: hidden;
 }
