+++ assets/js/backtotop.js
@@ -1,0 +1,14 @@
+function showBackToTop() {
+	let bt = document.getElementById("back-to-top");
+	bt.style.display = "inline";
+}
+
+function hideBackToTop() {
+	let bt = document.getElementById("back-to-top");
+	bt.style.display = "none";
+}
+
+export {
+	showBackToTop,
+	hideBackToTop,
+};
