--- assets/js/uistate.js
@@ -1,7 +1,0 @@
-﻿let uiState = {};
-uiState.hasMenu = true;
-uiState.hasContactModal = false;
-uiState.hasPrivacyModal = false;
-uiState.hasBackToTop = false;
-
-export {uiState};
