+++ index.html
@@ -23,7 +23,7 @@
 
 <main>
 	<header id="header">
-		<h1 class="logo">Team Xerbutri</h1>
+		<h1>Team Xerbutri</h1>
 	</header>
 
 	<div id="app">
