+++ main.js
@@ -34,11 +34,7 @@ let navState = new stateContext();
 	}
 
 	function init() {
-		// load image
-		document.getElementById("tx").src = `${txLogo}`;
-
 		manipulateHref();
-
 		navState.initState();
 	}
 })();
